Fix for blood-belt-express recipe
This one was a head scratcher. The recipe for express belts has an expensive variant. The code as written did not make the alteration correctly to modify that recipe, leaving one with no express blood belt recipe and duplicate express belt recipes.

The original code got around this I suspect by doing a deepcopy on the fast-belt recipe and avoiding the issue. As my Lua is not nearly swift enough (especially at 2:30 AM!) to figure out how to modify the normal/expensive recipe, I instead do this simple-ish tweak to handle the express case in a similar vein to the original implementation.

A quick test on my end shows it working, but additional eyes would always be beneficial, because 2 AM.